*** empty log message ***
git-svn-id: https://powerman.googlecode.com/svn/trunk@687 76eff002-5ed3-810f-47aa-be2f003abbef

+++ src/powerman.c
@@ -446,7 +446,7 @@ static void _getline(char *buf, int size)
 static int _process_line(void)
 {
     char buf[CP_LINEMAX];
-    int num;
+    long int num;
 
     _getline(buf, CP_LINEMAX);
